After ahsun work

+++ WindowsFormsApp3/Ledger.cs
@@ -72,12 +72,12 @@ namespace WindowsFormsApp3
             {
                 row = tempDataTable.NewRow();
                 row["Name"] = listView1.Items[i].SubItems[0].Text;
-                row["Amount"] = Convert.ToInt32(listView1.Items[i].SubItems[1].Text);
-                row["Received"] = Convert.ToInt32(listView1.Items[i].SubItems[2].Text);
-                row["Balance"] = Convert.ToInt32(listView1.Items[i].SubItems[3].Text);
-                row["TotalAmount"] = Convert.ToInt32(total_amount.Text);
-                row["TotalReceived"] = Convert.ToInt32(total_Rec.Text);
-                row["TotalBalance"] = Convert.ToInt32(total_balance.Text);
+                row["Amount"] = Convert.ToDouble(listView1.Items[i].SubItems[1].Text);
+                row["Received"] = Convert.ToDouble(listView1.Items[i].SubItems[2].Text);
+                row["Balance"] = Convert.ToDouble(listView1.Items[i].SubItems[3].Text);
+                row["TotalAmount"] = Convert.ToDouble(total_amount.Text);
+                row["TotalReceived"] = Convert.ToDouble(total_Rec.Text);
+                row["TotalBalance"] = Convert.ToDouble(total_balance.Text);
                 tempDataTable.Rows.Add(row);
             }
